requirements.txt
PyYAML
numpy
pandas
torch
transformers
python-telegram-bot
